council meeting display adjustments and form fixes

+++ frontend/controllers/AgendaController.php
@@ -7,9 +7,11 @@ use frontend\models\Agenda;
 use frontend\models\AgendaMinutes;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\ForbiddenHttpException;
 use yii\bootstrap4\ActiveForm;
 use yii\filters\VerbFilter;
 use yii\helpers\Html;
+use frontend\models\Audit;
 
 /**
  * AgendaController implements the CRUD actions for Agenda model.
@@ -73,13 +75,18 @@ class AgendaController extends Controller
                 'agenda.date',
                 'agenda.body AS aBody',
                 'agenda.create_dt AS aCreateDt',
+                'agenda.created_by AS aCreatedBy',
+                'user.first_name As uFname',
+                'user.last_name As uLname',
                 'agenda_minutes.id AS mId',
                 'agenda_minutes.attend',
                 'agenda_minutes.absent',
                 'agenda_minutes.body AS mBody',
+                'agenda_minutes.video AS mVideo',
                 'agenda_minutes.create_dt AS mCreateDt'
             ])
             ->leftJoin('agenda_minutes', '`agenda_minutes`.`agenda_id` = `agenda`.`id`')
+            ->leftJoin('user', '`user`.`id` = `agenda`.`created_by`')
             ->where(['agenda.id'=>$id])->asArray()->one();
             //$model = AgendaMinutes::find()->select([
         //        'agenda_minutes.*', 'agenda.*'
@@ -115,26 +122,37 @@ class AgendaController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Agenda();
-        $model->scenario ='create';   //helps with unique meeting on date validation
+        if (Yii::$app->user->can('create_agenda')) {
+            $model = new Agenda();
+            $model->scenario ='create';   //helps with unique meeting on date validation
 
-        if ($model->load(Yii::$app->request->post())) {
+            if ($model->load(Yii::$app->request->post())) {
 
-            $model->create_dt = date('Y-m-d');
-            $dateNote = date("l F jS", strtotime($model->date));
-            $model->date = date("Y-m-d", strtotime($model->date));
-            
-            if ($model->save()) {
-                Yii::$app->session->setFlash('success', "Agenda successfully set for " . $dateNote . " meeting.");
-            } else {
-                Yii::$app->session->setFlash('error', "Failed to set meeting agenda. Error: " . Html::error($model,'date'));
+                $model->create_dt = date('Y-m-d');
+                $model->created_by = Yii::$app->user->identity->id;
+                $dateNote = date("l F jS", strtotime($model->date));
+                $model->date = date("Y-m-d", strtotime($model->date));
+                
+                if ($model->save()) {
+                    $audit = new Audit();
+                    $audit->table = 'agenda';
+                    $audit->record_id = $model->id;
+                    $audit->field = 'Create';
+                    $audit->update_user = $model->created_by;
+                    $audit->save(false);
+                    Yii::$app->session->setFlash('success', "Agenda successfully set for " . $dateNote . " meeting.");
+                } else {
+                    Yii::$app->session->setFlash('error', "Failed to set meeting agenda. Error: " . Html::error($model,'date'));
+                }
+                return $this->redirect(['sibley/council', 'id' => $model->id]);
             }
-            return $this->redirect(['sibley/council', 'id' => $model->id]);
-        }
 
-        return $this->renderAjax('create', [
-            'model' => $model,
-        ]);
+            return $this->renderAjax('create', [
+                'model' => $model,
+            ]);
+        } else {
+            throw new ForbiddenHttpException('You do not have permission to perform this action.');
+        }    
     }
 
     /**
@@ -146,24 +164,42 @@ class AgendaController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (Yii::$app->user->can('update_agenda')) {
 
-        if ($model->load(Yii::$app->request->post())) {
-            $dateNote = date("l F jS", strtotime($model->date));
-            $model->date = date("Y-m-d", strtotime($model->date));
-            
-            if ($model->save()) {
-                Yii::$app->session->setFlash('success', "Agenda successfully updated for $dateNote meeting.");
-            } else {
-                Yii::$app->session->setFlash('error', "Failed to update meeting agenda. Error: " . Html::error($model,'date'));
+            $model = $this->findModel($id);
+
+            //make sure only owner or site admin can update
+            $user_id = Yii::$app->user->identity->id;
+            if ($user_id != $model->created_by && $user_id != 1) {
+                Yii::$app->session->setFlash('error', "It does not appear you created this agenda. Update request rejected.");
+                return $this->goBack(Yii::$app->request->referrer);
             }
-            return $this->redirect(['sibley/council', 'id' => $model->id]);
-        }
 
-        $model->date = date("m/d/Y", strtotime($model->date));
-        return $this->renderAjax('update', [
-            'model' => $model,
-        ]);
+            if ($model->load(Yii::$app->request->post())) {
+                $dateNote = date("l F jS", strtotime($model->date));
+                $model->date = date("Y-m-d", strtotime($model->date));
+                
+                if ($model->save()) {
+                    $audit = new Audit();
+                    $audit->table = 'agenda';
+                    $audit->record_id = $model->id;
+                    $audit->field = 'Update';
+                    $audit->update_user = Yii::$app->user->identity->id;
+                    $audit->save(false);
+                    Yii::$app->session->setFlash('success', "Agenda successfully updated for $dateNote meeting.");
+                } else {
+                    Yii::$app->session->setFlash('error', "Failed to update meeting agenda. Error: " . Html::error($model,'date'));
+                }
+                return $this->redirect(['sibley/council', 'id' => $model->id]);
+            }
+
+            $model->date = date("m/d/Y", strtotime($model->date));
+            return $this->renderAjax('update', [
+                'model' => $model,
+            ]);
+        } else {
+            throw new ForbiddenHttpException('You do not have permission to perform this action.');
+        }
     }
 
     /**
@@ -175,13 +211,32 @@ class AgendaController extends Controller
      */
     public function actionDelete($id)
     {
-        if ($this->findModel($id)->delete()) {
-            Yii::$app->session->setFlash('success', "Agenda successfully deleted.");
-        } else {
-            Yii::$app->session->setFlash('error', "An error occured while deleting this agenda");
-        }
+        if (Yii::$app->user->can('delete_agenda')) {
 
-        return $this->redirect(['sibley/council']);
+            //make sure only owner or site admin can update
+            $user_id = Yii::$app->user->identity->id;
+            if ($user_id != $model->created_by && $user_id != 1) {
+                Yii::$app->session->setFlash('error', "It does not appear you created this agenda. Delete request rejected.");
+                return $this->goBack(Yii::$app->request->referrer);
+            }
+        
+            if ($this->findModel($id)->delete()) {
+                $audit = new Audit();
+                $audit->table = 'agenda';
+                $audit->record_id = $id;
+                $audit->field = 'Delete';
+                $audit->update_user = Yii::$app->user->identity->id;
+                $audit->save(false);
+                Yii::$app->session->setFlash('success', "Agenda successfully deleted.");
+            } else {
+                Yii::$app->session->setFlash('error', "An error occured while deleting this agenda");
+            }
+
+            return $this->redirect(['sibley/council']);
+
+        } else {
+            throw new ForbiddenHttpException('You do not have permission to perform this action.');
+        }
     }
 
     /**
